feat(openclaw): expose Algorand wallet address
