+++ packages/ac2-open-claw-reference/README.md
@@ -10,7 +10,7 @@ Liquid Auth + WebRTC via [`@algorandfoundation/ac2-sdk`](../ac2-sdk).
 | OpenClaw surface        | AC2 contribution                                                           |
 | ----------------------- | -------------------------------------------------------------------------- |
 | Channel `ac2`           | Owns Liquid Auth + WebRTC pairing and the active session.                  |
-| Tool `ac2_capabilities` | Agent DID + `sig_hint` catalog.                                            |
+| Tool `ac2_capabilities` | Agent DID, connected wallet address, and `sig_hint` catalog.              |
 | Tool `ac2_sign`         | Routes a `SigningRequest` and returns signature details to the agent.      |
 | Tool `ac2_x402_fetch`   | Pays x402 exact Algorand resources using wallet-approved AC2 signing.      |
 | Setup entry             | `openclaw ac2 setup` writes the channel/tools wiring into `openclaw.json`. |
